Test Calculator

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Application.java
@@ -3,9 +3,9 @@ package com.kodilla.basic_assertion;
 public class Application {
     public static void main(String[] args) {
         Calculator calculator = new Calculator();
-        int a = 10;
-        int b = 30;
-        int addResult = calculator.add(a,b);
+        double a = 10;
+        double b = 30;
+        double addResult = calculator.add(a,b);
         boolean addCorrect = ResultChecker.assertEquals(40, addResult);
         if(addCorrect){
             System.out.println("Metoda add dziala poprawnie dla liczb: " + a + " i " + b);
@@ -13,7 +13,7 @@ public class Application {
             System.out.println("Metoda add dziala niepoprawnie dla liczb: " + a + " i " + b);
         }
 
-        int subResult = calculator.sub(a,b);
+        double subResult = calculator.sub(a,b);
         boolean subCorrect = ResultChecker.assertEquals(-20,subResult);
         if (subCorrect){
             System.out.println("Metoda sub dziala poprawnie dla liczb: " + a + " i " + b);
@@ -21,7 +21,7 @@ public class Application {
             System.out.println("Metoda sub nie dziala poprawnie dla liczb: " + a + " i " + b);
         }
 
-        int squareResult = calculator.square(a);
+        double squareResult = calculator.square(a);
         boolean squareCorrect = ResultChecker.assertEquals(100, squareResult);
         if(squareCorrect){
             System.out.println("Metoda square dziala poprawnie dla liczby: " + a);
