+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -2,15 +2,15 @@ package com.kodilla.basic_assertion;
 
 public class Calculator {
 
-    public int add(int a, int b){
+    public double add(double a, double b){
         return a+b;
     }
 
-    public int sub(int a, int b){
+    public double sub(double a, double b){
         return  a-b;
     }
 
-    public int square(int a){
+    public double square(double a){
         return a*a;
     }
 }
